% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposr-package.R
\docType{package}
\name{exposr}
\alias{exposr-package}
\alias{exposr}
\title{exposr: A Pharmacometric Utility Toolkit}
\description{
Tools for simulating and visualizing dose–exposure–response relationships.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TheoPapath/exposr}
  \item Report bugs at \url{https://github.com/TheoPapath/exposr/issues}
}

}
\author{
\strong{Maintainer}: Theodoros Papathanasiou \email{theo.x.papathanasiou@gsk.com}

}
