% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposr_palette_1.R
\name{exposr_palette_1}
\alias{exposr_palette_1}
\title{rxposr palette for up to 8 groups}
\usage{
exposr_palette_1(n = 8)
}
\arguments{
\item{n}{Integer. Number of colors to return (max 8).}
}
\value{
Character vector of hex color codes.
}
\description{
Returns a character vector of up to 8 hex colors
}
