% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expos.R
\name{expos_model}
\alias{expos_model}
\title{Modeling Functions}
\usage{
expos_model(wind_direction, inflection_angle, save = TRUE, console = TRUE)
}
\arguments{
\item{wind_direction}{wind direction (degrees)}

\item{inflection_angle}{inflection angle (degrees)}

\item{save}{whether to save results to file}

\item{console}{whether to display messages in console}
}
\value{
no return value
}
\description{
expos_model uses a raster file of elevation values, a specified wind
direction, and a specified inflection angle to create a raster file
of wind exposure values (0 = missing data, 1 = protected, 2 = exposed).
}
