% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expos.R
\name{expos_plot}
\alias{expos_plot}
\title{Plotting Functions}
\usage{
expos_plot(filename)
}
\arguments{
\item{filename}{name of input raster file}
}
\value{
no return value
}
\description{
expos_plot creates a plot of a specified raster file.
}
