% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expos.R
\name{expos_summarize}
\alias{expos_summarize}
\title{Summarizing Functions}
\usage{
expos_summarize(filename, console = TRUE)
}
\arguments{
\item{filename}{name of input raster file}

\item{console}{whether to display results in console}
}
\value{
a string containing summary information
}
\description{
expos_summarize displays summary information for a specified raster
file, including the number of rows and columns, spatial extent, cell
height and width, and minimum and maximum value.
}
