% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expos.R
\name{expos_set_path}
\alias{expos_set_path}
\title{Utility Functions}
\usage{
expos_set_path(exp_path, console = TRUE)
}
\arguments{
\item{exp_path}{path for current model runs}

\item{console}{whether to display messages in console}
}
\value{
no return value
}
\description{
expos_set_path sets the path for the current set of model runs.
}
