% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expos.R
\name{expos_model}
\alias{expos_model}
\alias{expos_damage}
\title{Modeling Functions}
\usage{
expos_model(
  wind_direction,
  inflection_angle,
  lat_long = FALSE,
  orient = 0,
  save = TRUE,
  console = TRUE
)

expos_damage(hurricane, inflection_angle, protect, save = TRUE, console = TRUE)
}
\arguments{
\item{wind_direction}{wind direction (degrees)}

\item{inflection_angle}{inflection angle (degrees)}

\item{lat_long}{whether coordinate system is latitude/longitude (degrees)}

\item{orient}{map orientation (degrees)}

\item{save}{whether to save results to file}

\item{console}{whether to display messages in console}

\item{hurricane}{hurricane name (as it appears in tif file)}

\item{protect}{how much to reduce damage in protected areas (Fujita 
scale ratings)}
}
\value{
raster of modeled exposure values

raster of modeled wind damage values
}
\description{
expos_model uses a raster file of elevation values, a specified wind
direction, and a specified inflection angle to create a raster file
of wind exposure values (0 = missing data, 1 = protected, 2 = exposed).
If a geographic coordinate system is used, horizontal and vertical units 
are assumed to be degrees and meters, respectively; otherwise horizontal 
and vertical units must be the same. Columns are assumed to be closely 
aligned with true North (0 degrees); if not, the map orientation must 
be specified. The name of the input file is assumed to be "dem.tif".

expos_damage uses output from Hurrecon and Expos to create a raster
of hurricane wind damage where topograhic exposure at each location
is determined by peak wind direction. If a location is protected, 
the enhanced Fujita scale rating is reduced by a specified amount.
This function requires a hurricane tif file created by Hurrecon, 
eight exposure files created by Expos (N, NE, E, etc), and a reprojection
file in csv format that contains lat long coordinates for the lower left 
and upper right corners of the digital elevation model.
}
