% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expos.R
\name{expos_plot}
\alias{expos_plot}
\title{Plotting Functions}
\usage{
expos_plot(
  filename,
  title = "",
  h_units = "meters",
  v_units = "meters",
  vector = TRUE,
  colormap = "default"
)
}
\arguments{
\item{filename}{name of input raster file}

\item{title}{plot title}

\item{h_units}{horizontal units}

\item{v_units}{vertical units}

\item{vector}{whether to display vectory boundary files}

\item{colormap}{color palette}
}
\value{
no return value
}
\description{
expos_plot creates a plot of a specified raster file. Optional arguments
include plot title, horizontal units, vertical units, vector (boundary
files) and color palette.
}
