% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expos.R
\name{expos_plot}
\alias{expos_plot}
\title{Plotting Functions}
\usage{
expos_plot(filename, title = "", col = rev(terrain.colors(255)))
}
\arguments{
\item{filename}{name of input raster file}

\item{title}{plot title}

\item{col}{color palette}
}
\value{
no return value
}
\description{
expos_plot creates a plot of a specified raster file. Optional arguments
include plot title and color palette.
}
