% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expos.R
\name{expos_set_path}
\alias{expos_set_path}
\alias{expos_get_path}
\title{Utility Functions}
\usage{
expos_set_path(exp_path)

expos_get_path()
}
\arguments{
\item{exp_path}{path for current model runs}
}
\value{
no return value

current path
}
\description{
expos_set_path sets the path for the current set of model runs.

expos_get_path returns the current path for a set of model runs.
}
